samples: drivers: espi: Refactor sample code to ease extensions
CAF eSPI flash API sample code and documentation about EC region
TAF eSPI APIs sample code
